// src/components/acl-table-pagination/acl-table-pagination.type.tsx
import { PaginationProps } from '@mui/material';
import React from 'react';

export declare type AclTablePaginationProps = PaginationProps & {
  children?: React.ReactNode;
  rowsPerPage?: number[];
  defaultRowsPerPage?: number;
  totalNumberOfRows?: number;
  rowsPerPageValue?: number;
  onChangeRowsPerPage?: (event: number) => void;
  onChangePage?: (event: number) => void;
};